implement search function, refactor some code

+++ src/hilbert.h
@@ -1,4 +1,4 @@
 #include "rtree.h"
 #define N 20
 
-int hilbertValue(POINT p);
+float hilbertValue(POINT p);
